refactor(landing): update imports to use index files for better organization
