+++ apps/demo/tsconfig.app.json
@@ -29,11 +29,11 @@
     "../../libs/shell/src/lib/nav/NavMain.tsx",
     "../../libs/shell/src/lib/nav/NavUser.tsx",
     "../../libs/shell/src/lib/nav/TeamSwitcher.tsx",
-    "../../libs/landing/src/lib/components/about-section/AboutSection.tsx",
+    "../../libs/landing/src/lib/components/about-section/index.tsx",
     "../../libs/landing/src/lib/components/features-section/FeaturesSection.tsx",
-    "../../libs/landing/src/lib/components/footer-section/Footer.tsx",
+    "../../libs/landing/src/lib/components/footer-section/index.tsx",
     "../../libs/landing/src/lib/components/get-started-section/StepsSection.tsx",
-    "../../libs/landing/src/lib/components/hero-section/HeroSection.tsx",
-    "../../libs/landing/src/lib/components/steps-section/StepsSection.tsx"
+    "../../libs/landing/src/lib/components/hero-section/index.tsx",
+    "../../libs/landing/src/lib/components/steps-section/index.tsx"
   ]
 }
