+++ libs/landing/src/lib/components/about-section/AboutSection.stories.tsx
@@ -23,7 +23,7 @@
  */
 
 import type { Meta, StoryObj } from '@storybook/react';
-import { AboutSection } from './AboutSection';
+import { AboutSection } from '.';
 
 
 const meta: Meta<typeof AboutSection> = {
