+++ libs/landing/src/lib/components/about-section/index.tsx
