+++ libs/landing/src/lib/components/footer-section/index.tsx
