+++ libs/landing/src/lib/components/hero-section/index.tsx
