+++ libs/landing/src/lib/components/index.ts
@@ -1,6 +1,6 @@
-export * from './about-section/AboutSection';
-export * from './cta-section/CTASection';
+export * from './about-section';
+export * from './cta-section';
 export * from './features-section';
-export * from './footer-section/Footer';
-export * from './steps-section/StepsSection';
-export * from './hero-section/HeroSection';
+export * from './footer-section';
+export * from './steps-section';
+export * from './hero-section';
