+++ libs/landing/src/lib/components/steps-section/index.tsx
